fix(mail): clarify rule action display
