+++ apps/web/app/(app)/[emailAccountId]/assistant/ResultDisplay.tsx
@@ -18,12 +18,7 @@ import { useRuleDialog } from "@/app/(app)/[emailAccountId]/assistant/RuleDialog
 import { ThreadSkipHint } from "@/app/(app)/[emailAccountId]/assistant/ThreadSkipHint";
 import { LearnedPatternExclusionHint } from "@/app/(app)/[emailAccountId]/assistant/LearnedPatternExclusionHint";
 import type { RunRulesResult } from "@/utils/ai/choose-rule/run-rules";
-import {
-  getActionDisplay,
-  getActionIcon,
-  getVisibleActions,
-} from "@/utils/action-display";
-import { getActionColor } from "@/components/PlanBadge";
+import { RuleActions } from "@/components/RuleActions";
 import { useAccount } from "@/providers/EmailAccountProvider";
 
 export function ResultsDisplay({
@@ -146,7 +141,7 @@ export function ResultDisplayContent({ result }: { result: RunRulesResult }) {
         {result.actionItems?.length ? (
           <>
             <div className="font-medium text-sm mb-1">Actions:</div>
-            <Actions
+            <RuleActions
               actions={
                 result.actionItems?.map((action) => ({
                   id: action.id,
@@ -211,69 +206,6 @@ export function ResultDisplayContent({ result }: { result: RunRulesResult }) {
       )}
 
       <RuleDialogComponent />
-    </div>
-  );
-}
-
-function Actions({
-  actions,
-  provider,
-  labels,
-}: {
-  actions: {
-    id: string;
-    type: ActionType;
-    label?: string | null;
-    labelId?: string | null;
-    folderName?: string | null;
-    content?: string | null;
-    to?: string | null;
-    subject?: string | null;
-    cc?: string | null;
-    bcc?: string | null;
-    url?: string | null;
-  }[];
-  provider: string;
-  labels: Array<{ id: string; name: string }>;
-}) {
-  return (
-    <div className="flex flex-col gap-2 flex-wrap">
-      {getVisibleActions(actions).map((action) => {
-        const Icon = getActionIcon(action.type);
-        const fields = [
-          { key: "to", value: action.to },
-          { key: "cc", value: action.cc },
-          { key: "bcc", value: action.bcc },
-          { key: "subject", value: action.subject },
-          { key: "content", value: action.content },
-          { key: "url", value: action.url },
-        ].filter((field) => field.value);
-
-        return (
-          <div key={action.id} className="flex flex-col gap-1">
-            <Badge
-              color={getActionColor(action.type)}
-              className="w-fit text-nowrap"
-            >
-              <Icon className="size-3 mr-1.5" />
-              {getActionDisplay(action, provider, labels)}
-            </Badge>
-            {fields.length > 0 && (
-              <div className="ml-1 space-y-0.5 text-sm text-muted-foreground">
-                {fields.map((field) => (
-                  <div
-                    key={field.key}
-                    className="whitespace-pre-wrap break-all"
-                  >
-                    <span className="font-medium capitalize">{field.key}:</span>{" "}
-                    {field.value}
-                  </div>
-                ))}
-              </div>
-            )}
-          </div>
-        );
-      })}
     </div>
   );
 }
